added base_headers class
